+++ CHANGELOG.md
@@ -5,7 +5,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
 ## Unreleased
 
 ### Changed
@@ -24,6 +23,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Remove the following deprecated methods from Uuid classes:
   * `getClockSeqHiAndReservedHex()`
 * Remove dependency on ramsey/collection package.
+
+
+## 4.3.1 - 2022-03-27
 
 ### Deprecated